[Task-37] implemented mechanism for displaying icon status and also small refactor has been done
