+++ BudgetTracker/BudgetTracker/src/app/ui/components/modals/budgets-modal/budgets-modal.component.html
@@ -14,7 +14,7 @@
     <form #form="ngForm">
       @for (budgetField of budgetDateFields; let i = $index; track i) {
         <div class="d-flex align-items-baseline">
-          @if (!isUndefinedStatus(budgetResponses[i])) {
+          @if (!ModalUtils.isUndefinedBudgetStatus(budgetResponses[i])) {
             @if (budgetResponses[i].status) {
               <div class="text-success"
                    data-bs-toggle="tooltip"
@@ -50,7 +50,7 @@
               (ngModelChange)="validateMonthsOccurs(i, toDate)"
               #dateEnd="ngbDatepicker"
               [ngClass]="{
-                    'is-invalid': toDate.invalid || (!isUndefinedStatus(budgetResponses[i]) && !budgetResponses[i].status),
+                    'is-invalid': toDate.invalid || (!ModalUtils.isUndefinedBudgetStatus(budgetResponses[i]) && !budgetResponses[i].status),
                     'is-valid': toDate.valid }"/>
             <div class="input-group-append">
               <button class="btn btn-outline-secondary bi bi-calendar3"
@@ -85,7 +85,7 @@
           <div>
             <button type="button"
                     [disabled]="disableForm ||
-                     (!isUndefinedStatus(budgetResponses[i]) && budgetResponses[i].status)"
+                     (!ModalUtils.isUndefinedBudgetStatus(budgetResponses[i]) && budgetResponses[i].status)"
                     class="btn btn-outline-danger"
                     data-bs-toggle="tooltip"
                     data-bs-placement="bottom"
