+++ BudgetTracker/BudgetTracker/src/app/ui/components/modals/budgets-modal/budgets-modal.component.ts
@@ -1,13 +1,14 @@
 import {Component, EventEmitter, OnDestroy, OnInit, Output, ViewChild} from '@angular/core';
 import {NgbModal} from "@ng-bootstrap/ng-bootstrap";
-import {ModalOptions} from "../../../../util/modal.utils";
+import {ModalOptions} from "../../../../util/modal-options.utils";
 import {DatePickerModel} from "../../../../models/FormModels";
 import {DateUtils} from "../../../../util/date.utils";
 import {HttpService} from "../../../../services/http/httpService";
 import {catchError, forkJoin, interval, Observable, of, Subscription, takeWhile} from "rxjs";
 import {HttpResponse} from "@angular/common/http";
 import {SubscriptionUtils} from "../../../../util/subscription.utils";
-import {BudgetStatus} from "../../../../models/BudgetStatusModel";
+import {BudgetStatus} from "../../../../models/modal-models/BudgetStatusModel";
+import {ModalUtils} from "../../../../util/modal.utils";
 
 
 @Component({
@@ -116,7 +117,7 @@ export class BudgetsModalComponent implements OnInit, OnDestroy {
       const field = this.budgetDateFields[i];
       const formatedDate = DateUtils.formatDatePicker(field);
       if (!this.budgetResponses[i].status ||
-        this.isUndefinedStatus(this.budgetResponses[i])) {
+        ModalUtils.isUndefinedBudgetStatus(this.budgetResponses[i])) {
         budgetRequests.push(this.httpService.createBudget(formatedDate).pipe(
           catchError((err): Observable<HttpResponse<any>> => {
             return of(err);
@@ -150,7 +151,7 @@ export class BudgetsModalComponent implements OnInit, OnDestroy {
         complete: (): void => {
           this.autoCloseModal = true;
           this.budgetResponses.forEach((value): void => {
-            if (!this.isUndefinedStatus(value) && !value.status) {
+            if (!ModalUtils.isUndefinedBudgetStatus(value) && !value.status) {
               this.autoCloseModal = false;
             }
           });
@@ -166,7 +167,7 @@ export class BudgetsModalComponent implements OnInit, OnDestroy {
   protected close(modal: any): void {
     let pageReload = false;
     for (const budgetResponse of this.budgetResponses) {
-      if (!this.isUndefinedStatus(budgetResponse)) {
+      if (!ModalUtils.isUndefinedBudgetStatus(budgetResponse)) {
         pageReload = true;
       }
     }
@@ -177,11 +178,6 @@ export class BudgetsModalComponent implements OnInit, OnDestroy {
     }
 
     modal.close();
-  }
-
-  protected isUndefinedStatus(budgetStatus: BudgetStatus): boolean {
-    return typeof budgetStatus.status === "undefined" &&
-      typeof budgetStatus.message === "undefined"
   }
 
   protected startTimer(): void {
@@ -222,4 +218,6 @@ export class BudgetsModalComponent implements OnInit, OnDestroy {
       message: err.status + " - " + err.error["title"]
     } as BudgetStatus;
   }
+
+  protected readonly ModalUtils = ModalUtils;
 }
