+++ BudgetTracker/BudgetTracker/src/app/util/modal.utils.ts
@@ -1,36 +1,8 @@
-import { NgbModalOptions } from '@ng-bootstrap/ng-bootstrap/modal/modal-config';
+import {BudgetStatus} from "../models/modal-models/BudgetStatusModel";
 
-export class ModalOptions {
-
-  static default(ms?: ModalSize): NgbModalOptions {
-    let size: string;
-    switch (ms) {
-      case ModalSize.SMALL:
-        size = 'sm';
-        break;
-      case ModalSize.MEDIUM:
-        size = 'md';
-        break;
-      case ModalSize.BIG:
-        size = 'lg';
-        break;
-      default:
-        size = 'md';
-        break;
-    }
-
-    return {
-      centered: true,
-      scrollable: true,
-      size: size,
-      backdrop: 'static',
-      keyboard: false,
-    } as NgbModalOptions;
+export class ModalUtils {
+  public static isUndefinedBudgetStatus(budgetStatus: BudgetStatus): boolean {
+    return typeof budgetStatus.status === "undefined" &&
+      typeof budgetStatus.message === "undefined"
   }
-}
-
-export enum ModalSize {
-  SMALL,
-  MEDIUM,
-  BIG,
 }
